docs: correct wrong import of PS inside JS in FFI.md
The `… from 'Test'` import doesn't work unless some additional setup is done. Presumably this is the hint about `psc` compiler, but the documentation gives no detail on such configuration as if it would Just Work™. Well, it doesn't work — what does, however, is using relative path. It works well with both `spago run` and `spago bundle-app` (barring the fact that `gcd` has name clash in Prelude).

So replace the path to make sure it works with the default setup people most likely have.